refactor glossary imports
